+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 ### Changes
 
 - Include link and size info to the full size image in `ImageView` view component @sneridagh
+- Hide Blocks fields in Layout fieldset in Add/Edit forms @pnicolli
+
+### Internal
+
+- Added forest.eea.europa.eu as deployed Volto in production @tiberiuichim
 
 ## 4.0.0-alpha.36 (2020-02-03)